[velox] refactor(functions): Move Spark decimal arithmetic structs to header
GPU SFI needs to compile Spark decimal arithmetic `call()` bodies from device TUs. Eleven structs (including `DecimalAddSubtractBase` and `DecimalIntegralDivideBase`) lived in an anonymous namespace in `DecimalArithmetic.cpp`, so no other translation unit could `#include` them.

Move the structs verbatim into `DecimalArithmeticFunctions.h` under `facebook::velox::functions::sparksql`. Keep `kDenyPrecisionLoss`, template aliases, and registration helpers in `DecimalArithmetic.cpp`. No logic or registration API changes.

Issue: https://github.com/facebookincubator/velox/issues/17259

Test plan:
- Build and run `velox_functions_spark_test --gtest_filter=DecimalArithmeticTest.*` in the container
