+++ velox/functions/sparksql/CMakeLists.txt
@@ -72,6 +72,7 @@ velox_add_library(
   ConcatWs.h
   DateTimeFunctions.h
   DecimalArithmetic.h
+  DecimalArithmeticFunctions.h
   DecimalCeil.h
   DecimalUtil.h
   Factorial.h
